+++ src/auth/login-with-extension.ts
@@ -29,18 +29,16 @@ export const loginWithExtension = async (
       if (address) {
         EventsStore.run(EventStoreEvents.onLoginStart);
       } else {
-        EventsStore.run(EventStoreEvents.onLoginFailure);
+        throw new Error('There were problems while logging in!');
       }
     }
   } catch (e) {
     const err = errorParse(e);
-    console.warn(`Something went wrong trying to login the user: ${err}`);
+    throw new Error(err);
   }
 
   if (!dappProvider) {
-    throw new Error(
-      'Error: There were problems with auth provider initialization!'
-    );
+    throw new Error('There were problems with auth provider initialization!');
   }
 
   const { signature } = dappProvider.account;
@@ -76,10 +74,9 @@ export const loginWithExtension = async (
 
       return dappProvider;
     } catch (e: any) {
-      console.warn(
+      throw new Error(
         `Something went wrong trying to synchronize the user account: ${e?.message}`
       );
-      EventsStore.run(EventStoreEvents.onLoginFailure);
     }
   }
 };