+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### [0.16.0](https://github.com/elven-js/elven.js/releases/tag/v0.16.0) (2024-01-17)
-- rename and add more login callbacks (breaking)
+### [0.16.0](https://github.com/elven-js/elven.js/releases/tag/v0.16.0) (2024-01-...)
+- rename and add more login callbacks (breaking) (check [docs](https://www.elvenjs.com/docs/sdk-reference.html#login))
+- add callbacks for logout (check [docs](https://www.elvenjs.com/docs/sdk-reference.html#logout))
+- add callbacks for `queryContract` (check [docs](https://www.elvenjs.com/docs/sdk-reference.html#querying-a-smart-contract))
 
 ### [0.15.0](https://github.com/elven-js/elven.js/releases/tag/v0.15.0) (2024-01-13)
 - add webview provider (based on sdk-dapp), required for xPortal Hub integration (experimental, need more tests and rewrites, it will probably land in a separate package in the following updates)